% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_code}
\alias{all_code}
\title{CODE of 2367 Airports/Locations.}
\format{A data frame with 2367 rows and 3 variables:
\describe{
  \item{State}{State }
  \item{Code}{code of airport/locations}
  \item{Name}{name of airport/locations }
}}
\source{
\url{https://w1.weather.gov/xml/current_obs/}
}
\usage{
all_code
}
\description{
A dataset containing the code and name of 2367 US airport/locations.
}
\keyword{datasets}
