% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obhistory.R
\name{obhistory}
\alias{obhistory}
\title{Read weather observation history of One Airport Location}
\usage{
obhistory(id)
}
\arguments{
\item{id}{one 4 character airport code}
}
\value{
a data frame with 7 variabes descripting the weather history of Weather and Temperature
}
\description{
This function read a table from an html page from National Weather Service website and output a dataframe
}
\examples{
obhistory("KAMW")
}
