% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_history.R
\name{plot_temp_history}
\alias{plot_temp_history}
\title{Plot 3 day (72 hours) temperature history of Selected Airport Location}
\usage{
plot_temp_history(id)
}
\arguments{
\item{id}{one 4 character airport code}
}
\description{
This function plot a time series of 3 day temperature history for Selected Airports
}
\examples{
plot_temp_history("KAMW")
}
